% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cantor_pair.R
\name{cantor_pair}
\alias{cantor_pair}
\title{Cantor Pairing Function}
\usage{
cantor_pair(xval)
}
\arguments{
\item{xval}{A numeric vector of length 2 containing the values to pair.}
}
\value{
A single numeric value representing the Cantor pairing of \code{xval}.
}
\description{
Computes the Cantor pairing value for a 2-element numeric vector.
}
\details{
The Cantor pairing function maps two non-negative integers (x, y)
to a single non-negative integer in a reversible way.
}
\examples{
\dontrun{
cantor_pair(c(3, 5))
}

}
