% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cantor_tuple.R
\name{cantor_tuple}
\alias{cantor_tuple}
\title{Cantor Tuple Function}
\usage{
cantor_tuple(xval)
}
\arguments{
\item{xval}{A numeric vector of length 1 or more.}
}
\value{
A single numeric value representing the Cantor-encoded tuple.
}
\description{
Extends the Cantor pairing function to tuples of arbitrary length.
}
\details{
This function reduces an n-element numeric vector to a single non-negative
integer using repeated Cantor pairing. For n = 1 the value is returned
unchanged; for n = 2 it behaves like \code{cantor_pair}; for larger n it applies
recursive pairing.
}
\examples{
\dontrun{
cantor_tuple(c(2, 4, 6))
}

}
\seealso{
cantor_pair
}
