% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_inv.R
\name{safe_inv}
\alias{safe_inv}
\title{Safely invert a matrix}
\usage{
safe_inv(matr)
}
\arguments{
\item{matr}{A numeric matrix to invert.}
}
\value{
The inverse of matr, or its Moore-Penrose pseudoinverse
if solve fails.
}
\description{
Tries to invert a matrix using the solve function. If \code{solve} fails,
usually because the matrix is singular or ill-conditioned,
the function returns the result of \code{mp_inv} instead.
}
\examples{
\dontrun{
safe_inv(diag(3))
}

}
\seealso{
mp_inv
}
